+++ apps/tc/theories/tc.v
@@ -1,7 +1,8 @@
 (* license: GNU Lesser General Public License Version 2.1 or later           *)
 (* ------------------------------------------------------------------------- *)
 
-Declare ML Module "rocq-elpi.tc".
+From elpi Require Import elpi.
+Declare ML Module "rocq-elpi-typeclasses.elpi_tc_plugin".
 
 From elpi.apps.tc.elpi Extra Dependency "tc_aux.elpi" as tc_aux.
 (* From elpi.apps.tc.elpi Extra Dependency "compiler.elpi" as compiler. *)